Avoid duplicating search patterns
The method mysqlSyntax at line 622 is returning a Collection that is the
same collection received as second the parameter "pattern". To avoid
duplicating entries in the Collection, must be called with a new
collection (or don't call addAll on the same collection).

The bug had no functional impact but it had performance impact. Any text
search query was searching twice on JackRabbit looking for all
attachments containing that text.